zt411-troubleshooter-agent/pyproject.toml
[tool.poetry]
name = "zt411-troubleshooter-agent"
version = "0.1.0"
description = "Agentic AI troubleshooting system for Zebra ZT411 printers"
authors = ["Dylan Figureredo"]
license = "MIT"
packages = [{ include = "zt411_agent", from = "src" }]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.110"
uvicorn = "^0.29"
pydantic = "^2.6"
sqlmodel = "^0.0.16"
redis = "^5.0"
rq = "^1.15"
sentence-transformers = "^2.6"
faiss-cpu = "^1.8"
mlflow = "^2.12"
httpx = "^0.27"
pysnmp = "^4.4"
pyyaml = "^6.0"
prometheus-client = "^0.20"
scikit-learn = "^1.4"
numpy = "^1.26"
pandas = "^2.2"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0"
pytest-cov = "^4.1"
ruff = "^0.3"
black = "^24.3"
isort = "^5.13"
pre-commit = "^3.6"

[tool.black]
line-length = 100

[tool.ruff]
line-length = 100
select = ["E", "F", "I"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
